update enter background function
