+++ team-soj/SojTeamGame/SojTeamGame/TapStarbucks/TapStarbucksViewController.swift
@@ -47,14 +47,10 @@ class TapStarbucksViewController: UIViewController {
             monsterIndex += 1
             scoreLabel.text = "分數: \(score)"
             
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateFormat = "yyyy年M月d日 H:m"
-            let dateString = dateFormatter.string(from: Date())
-            
             UserDefaults.standard.set(max(score, Record.current.gameOneScore), forKey: "gameOneScore")
             Record.current.addRecordGame(gameName: "tapStarbucks")
             Record.current.addRecordScore(gameScore: score)
-            Record.current.addRecordDate(gameDate: dateString)
+            Record.current.addRecordDate(gameDate: DateHelper.shared.convertDateFormat())
             coverView.isHidden = false
             tapStarbucksTimer.invalidate()
             gameEndLabel.text = "You Win!!"
@@ -138,15 +134,12 @@ class TapStarbucksViewController: UIViewController {
         } else {
             tapStarbucksTimer.invalidate()
             timeCount = monster.gameTime
-            
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateFormat = "yyyy年M月d日 H:m"
-            let dateString = dateFormatter.string(from: Date())
+            DateHelper.shared.convertDateFormat()
             
             UserDefaults.standard.set(max(score, Record.current.gameOneScore), forKey: "gameOneScore")
             Record.current.addRecordGame(gameName: "tapStarbucks")
             Record.current.addRecordScore(gameScore: score)
-            Record.current.addRecordDate(gameDate: dateString)
+            Record.current.addRecordDate(gameDate: DateHelper.shared.convertDateFormat())
             coverView.isHidden = false
             gameEndLabel.text = "時間到了！"
         }
